pyproject.toml
[project]
name = "auto-oiltrader"
version = "0.1.0"
description = "Autonomous CL/BZ futures and options trading system with continuous learning and human-in-the-loop safety gates"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  # Market data + execution
  "ibapi>=9.81",
  # Storage
  "duckdb>=1.0",
  "pyarrow>=15",
  # Numerics
  "numpy>=2.0",
  "scipy>=1.13",
  "pandas>=2.2",
  # Web / dashboard / api
  "streamlit>=1.40",
  "streamlit-autorefresh>=1.0",
  "plotly>=5.20",
  "matplotlib>=3.9",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  # Config
  "tomli>=2.0; python_version<'3.11'",
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "pydantic-settings>=2.5",
  # LLM
  "anthropic>=0.40",
  # News + scraping
  "feedparser>=6.0",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  # Notifications
  "requests>=2.32",
  # CLI
  "typer>=0.12",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "ruff>=0.7",
  "mypy>=1.11",
]

[project.scripts]
tradectl = "cli.tradectl:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "strategies", "daemons", "agents", "interfaces", "cli"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = false
